promote mirrored options from drishti-upload (CI project) to top level
A set of options required by the drishti-upload CI cache project is mirrored internally in drishti in cases where drishti-upload is skipped, as is the case when using drishti as a package through hunter.  These were previously in the defualt Hunter LOCAL config cmake/Hunter/config.cmake, which means they would be skipped when hunter is used from a top level project that provides its’ own LOCAL cmake/Hunter/config.cmake file.  A few of these options are central to the build logic, so they are added to the top level CMakeLists.txt in cases where drishti-upload is no used.

Note: This logic (in particular the need for a drishti-upload project for CI builds) will be simplified when the cache uploads are managed directly by hunter (soon!).

* mirror drishti-upload options in top level cmake when drishti-upload isn’t used
* add option to support printing all cmake variables inside hunter for debugging (default off)